+++ apps/meteor/client/lib/normalizeThreadTitle.ts
@@ -5,8 +5,8 @@ import { Meteor } from 'meteor/meteor';
 import { emojiParser } from '../../app/emoji/client/emojiParser';
 import { filterMarkdown } from '../../app/markdown/lib/markdown';
 import { MentionsParser } from '../../app/mentions/lib/MentionsParser';
-import { Users } from '../../app/models/client';
 import { settings } from '../../app/settings/client';
+import { Users } from '../stores';
 
 export function normalizeThreadTitle({ ...message }: Readonly<IMessage>) {
 	if (message.msg) {
@@ -15,7 +15,7 @@ export function normalizeThreadTitle({ ...message }: Readonly<IMessage>) {
 			return filteredMessage;
 		}
 		const uid = Meteor.userId();
-		const me = (uid && Users.findOne(uid, { fields: { username: 1 } })?.username) || '';
+		const me = (uid && Users.state.get(uid)?.username) || '';
 		const pattern = settings.get('UTF8_User_Names_Validation');
 		const useRealName = settings.get('UI_Use_Real_Name');