+++ apps/meteor/client/views/admin/invites/InvitesPage.tsx
@@ -1,12 +1,12 @@
 import { States, StatesIcon, StatesTitle, StatesActions, StatesAction } from '@rocket.chat/fuselage';
 import { useMediaQuery } from '@rocket.chat/fuselage-hooks';
+import { GenericModal } from '@rocket.chat/ui-client';
 import { useSetModal, useToastMessageDispatch, useTranslation, useEndpoint } from '@rocket.chat/ui-contexts';
 import { useQuery } from '@tanstack/react-query';
 import type { ReactElement } from 'react';
 import { useMemo } from 'react';
 
 import InviteRow from './InviteRow';
-import GenericModal from '../../../components/GenericModal';
 import GenericNoResults from '../../../components/GenericNoResults';
 import {
 	GenericTable,