+++ apps/meteor/client/views/room/RoomOpener.tsx
@@ -1,15 +1,12 @@
 import type { RoomType } from '@rocket.chat/core-typings';
 import { Box, States, StatesIcon, StatesSubtitle, StatesTitle } from '@rocket.chat/fuselage';
-import { FeaturePreviewOff, FeaturePreviewOn } from '@rocket.chat/ui-client';
 import type { ReactElement } from 'react';
 import { lazy, Suspense } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import NotSubscribedRoom from './NotSubscribedRoom';
 import RoomSkeleton from './RoomSkeleton';
-import RoomSidepanel from './Sidepanel/RoomSidepanel';
 import { useOpenRoom } from './hooks/useOpenRoom';
-import { FeaturePreviewSidePanelNavigation } from '../../components/FeaturePreviewSidePanelNavigation';
 import { Header } from '../../components/Header';
 import { getErrorMessage } from '../../lib/errorHandling';
 import { NotAuthorizedError } from '../../lib/errors/NotAuthorizedError';
@@ -28,23 +25,12 @@ type RoomOpenerProps = {
 	reference: string;
 };
 
-const isDirectOrOmnichannelRoom = (type: RoomType) => type === 'd' || type === 'l';
-
 const RoomOpener = ({ type, reference }: RoomOpenerProps): ReactElement => {
 	const { data, error, isSuccess, isError, isLoading } = useOpenRoom({ type, reference });
 	const { t } = useTranslation();
 
 	return (
 		<Box display='flex' w='full' h='full'>
-			{!isDirectOrOmnichannelRoom(type) && (
-				<FeaturePreviewSidePanelNavigation>
-					<FeaturePreviewOff>{null}</FeaturePreviewOff>
-					<FeaturePreviewOn>
-						<RoomSidepanel />
-					</FeaturePreviewOn>
-				</FeaturePreviewSidePanelNavigation>
-			)}
-
 			<Suspense fallback={<RoomSkeleton />}>
 				{isLoading && <RoomSkeleton />}
 				{isSuccess && (