+++ apps/meteor/client/views/room/modals/ReactionListModal/ReactionListModal.tsx
@@ -1,9 +1,9 @@
 import type { IMessage } from '@rocket.chat/core-typings';
+import { GenericModal } from '@rocket.chat/ui-client';
 import type { ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import Reactions from './Reactions';
-import GenericModal from '../../../../components/GenericModal';
 
 type ReactionListModalProps = {
 	reactions: Required<IMessage>['reactions'];