+++ apps/meteor/server/lib/OutboundMessageProvider.ts
@@ -3,22 +3,27 @@ import type {
 	IOutboundMessageProviders,
 	IOutboundPhoneMessageProvider,
 } from '@rocket.chat/apps-engine/definition/outboundComunication';
-
-interface IOutboundMessageProvider {
-	registerPhoneProvider(provider: IOutboundPhoneMessageProvider): void;
-	registerEmailProvider(provider: IOutboundEmailMessageProvider): void;
-	getOutboundMessageProviders(type?: 'phone' | 'email'): IOutboundMessageProviders[];
-	unregisterProvider(appId: string, providerType: string): void;
-}
+import type { ValidOutboundProvider, IOutboundProvider, IOutboundMessageProvider } from '@rocket.chat/core-typings';
 
 export class OutboundMessageProvider implements IOutboundMessageProvider {
-	private readonly outboundMessageProviders: Map<'phone' | 'email', IOutboundMessageProviders[]>;
+	private readonly outboundMessageProviders: Map<ValidOutboundProvider, IOutboundMessageProviders[]>;
 
 	constructor() {
 		this.outboundMessageProviders = new Map([
 			['phone', []],
 			['email', []],
 		]);
+	}
+
+	public findOneByProviderId(providerId: string) {
+		for (const providers of this.outboundMessageProviders.values()) {
+			for (const provider of providers) {
+				if (provider.appId === providerId) {
+					return provider;
+				}
+			}
+		}
+		return undefined;
 	}
 
 	public registerPhoneProvider(provider: IOutboundPhoneMessageProvider): void {
@@ -29,15 +34,28 @@ export class OutboundMessageProvider implements IOutboundMessageProvider {
 		this.outboundMessageProviders.set('email', [...(this.outboundMessageProviders.get('email') || []), provider]);
 	}
 
-	public getOutboundMessageProviders(type?: 'phone' | 'email'): IOutboundMessageProviders[] {
+	public getOutboundMessageProviders(type?: ValidOutboundProvider): IOutboundProvider[] {
 		if (type) {
-			return Array.from(this.outboundMessageProviders.get(type)?.values() || []);
+			return Array.from(this.outboundMessageProviders.get(type)?.values() || []).map((provider) => ({
+				providerId: provider.appId,
+				providerName: provider.name,
+				providerType: provider.type,
+				...(provider.documentationUrl && { documentationUrl: provider.documentationUrl }),
+				...(provider.supportsTemplates && { supportsTemplates: provider.supportsTemplates }),
+			}));
 		}
 
-		return Array.from(this.outboundMessageProviders.values()).flatMap((providers) => providers);
+		return Array.from(this.outboundMessageProviders.values())
+			.flatMap((providers) => providers)
+			.map((provider) => ({
+				providerId: provider.appId,
+				providerName: provider.name,
+				supportsTemplates: provider.supportsTemplates,
+				providerType: provider.type,
+			}));
 	}
 
-	public unregisterProvider(appId: string, providerType: 'phone' | 'email'): void {
+	public unregisterProvider(appId: string, providerType: ValidOutboundProvider): void {
 		const providers = this.outboundMessageProviders.get(providerType);
 		if (!providers) {
 			return;